Resturctured load balancers to a separate package
