+++ driver/src/main/java/org/neo4j/driver/internal/cluster/loadbalancing/RoundRobinArrayIndex.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.internal.cluster;
+package org.neo4j.driver.internal.cluster.loadbalancing;
 
 import java.util.concurrent.atomic.AtomicInteger;
