+++ driver/src/test/java/org/neo4j/driver/internal/RoutingDriverTest.java
@@ -29,8 +29,8 @@ import java.io.File;
 import java.util.Collections;
 import java.util.Map;
 
-import org.neo4j.driver.internal.cluster.LoadBalancer;
 import org.neo4j.driver.internal.cluster.RoutingSettings;
+import org.neo4j.driver.internal.cluster.loadbalancing.LoadBalancer;
 import org.neo4j.driver.internal.net.BoltServerAddress;
 import org.neo4j.driver.internal.retry.FixedRetryLogic;
 import org.neo4j.driver.internal.retry.RetryLogic;
